pyproject.toml
[tool.poetry]
name = "progen"
version = "0.0.36"
description = "Protein Generation (ProGen)"
authors = ["Phil Wang <lucidrains@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.7,<3.9"
jax = "^0.2.20"
jaxlib = "^0.1.71"
python-dotenv = "^0.19.0"
dm-haiku = "^0.0.4"
jmp = "^0.0.2"
omegaconf = "^2.1.1"
optax = "^0.0.9"
pyfaidx = "^0.6.2"
tensorflow = "^2.6.0"
tqdm = "^4.62.2"
wandb = "^0.12.1"
cloudpickle = "^1.6.0"
einops = "^0.3.2"
click = ">7.0,<8.0"
click-option-group = "^0.5.3"
google-cloud-storage = "^1.42.0"
humanize = "^3.11.0"
Jinja2 = "<3.0"
prefect = "^0.15.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
